refactor(frontend): introduce shared/ui structure for primitives
Move existing UI primitives (Button, Card, Badge, Skeleton, ErrorCard) from
src/components/ui to src/shared/ui/<Component>/ with per-component folders and
index barrels. Add a top-level barrel src/shared/ui/index.ts so feature code
can import primitives from a single path.

Reserve src/shared/form and src/shared/table for upcoming PRs (TanStack Form +
effect/Schema, TanStack Table). Each has a README explaining intent.

Update all 16 import sites in features/ and routes/ to use '@/shared/ui'.
Promote 'clsx' from a transitive dependency to an explicit one. Ignore the
TanStack Router generated route tree (src/routeTree.gen.ts) since it is
regenerated by 'tsr generate' on every typecheck/build.

Move docs/UI_REFRESH_PLAN.md from the frontend project into docs/ where the
rest of architecture docs live.

No behavior changes. Verified with:
  npm run typecheck  ok
  npm run lint       ok
  npm run build      ok
  npm test           33/34 pass; the single failing test exists on main and
                     will be addressed in PR 2 when DeleteRuleDialog is
                     migrated to Radix AlertDialog.

Refs: docs/ui-refresh-plan.md (PR 1)
