+++ service-mesh/frontend/service-mesh/package.json
@@ -11,16 +11,17 @@
     "test": "vitest run",
     "test:watch": "vitest",
     "test:ui": "vitest --ui",
-    "lint":     "eslint src",
+    "lint": "eslint src",
     "lint:fix": "eslint src --fix"
   },
   "dependencies": {
+    "@tanstack/query-async-storage-persister": "^5.62.0",
+    "@tanstack/query-sync-storage-persister": "^5.62.0",
     "@tanstack/react-query": "^5.62.0",
     "@tanstack/react-query-persist-client": "^5.62.0",
-    "@tanstack/query-sync-storage-persister": "^5.62.0",
-    "@tanstack/query-async-storage-persister": "^5.62.0",
-    "@tanstack/router-devtools": "^1.95.0",
     "@tanstack/react-router": "^1.95.0",
+    "@tanstack/router-devtools": "^1.95.0",
+    "clsx": "^2.1.1",
     "effect": "^3.12.0",
     "idb-keyval": "^6.2.1",
     "lucide-react": "^0.468.0",
@@ -29,24 +30,24 @@
     "zustand": "^5.0.2"
   },
   "devDependencies": {
+    "@eslint/js": "^10.0.1",
     "@tanstack/router-cli": "^1.95.0",
     "@tanstack/router-plugin": "^1.95.0",
+    "@testing-library/jest-dom": "^6.6.0",
+    "@testing-library/react": "^16.0.0",
+    "@testing-library/user-event": "^14.5.2",
     "@types/react": "^19.0.0",
     "@types/react-dom": "^19.0.0",
     "@vitejs/plugin-react": "^4.3.4",
-    "@testing-library/react": "^16.0.0",
-    "@testing-library/jest-dom": "^6.6.0",
-    "@testing-library/user-event": "^14.5.2",
     "@vitest/ui": "^2.0.0",
-    "jsdom": "^25.0.0",
-    "typescript": "~5.7.2",
-    "vite": "^6.0.5",
-    "vitest": "^2.0.0",
-    "@eslint/js": "^10.0.1",
     "eslint": "^10.2.1",
     "eslint-plugin-fp": "^2.3.0",
     "eslint-plugin-react-hooks": "^7.0.1",
-    "typescript-eslint": "^8.59.0"
+    "jsdom": "^25.0.0",
+    "typescript": "~5.7.2",
+    "typescript-eslint": "^8.59.0",
+    "vite": "^6.0.5",
+    "vitest": "^2.0.0"
   },
   "overrides": {
     "eslint-plugin-react-hooks": {
