+++ service-mesh/frontend/service-mesh/src/features/routing-rules/components/RuleFormModal/RuleFormModal.tsx
@@ -2,7 +2,7 @@ import { type ReactElement, useEffect, useRef, useCallback } from 'react'
 import { X } from 'lucide-react'
 import type { RoutingRule, RuleFormValues } from '../../model/types'
 import { useRuleForm } from '../../hooks/useRuleForm'
-import { Button } from '@/components/ui/button'
+import { Button } from '@/shared/ui'
 import { RuleNameField } from './RuleNameField'
 import { RuleMatchFields } from './RuleMatchFields'
 import { DestinationList } from '../DestinationList/DestinationList'
