+++ service-mesh/frontend/service-mesh/src/features/routing-rules/components/RulesTable/RulesTable.tsx
@@ -1,7 +1,7 @@
 import type { ReactElement } from 'react'
 import { useState } from 'react'
 import { Trash2, Pencil } from 'lucide-react'
-import { Button } from '@/components/ui/button'
+import { Button } from '@/shared/ui'
 import type { RoutingRule } from '../../model/types'
 import { DeleteRuleDialog } from '../DeleteRuleDialog/DeleteRuleDialog'
 import styles from './RulesTable.module.css'
