+++ service-mesh/frontend/service-mesh/src/features/services/ServiceDetailPage.tsx
@@ -3,8 +3,7 @@ import { useQuery } from '@tanstack/react-query'
 import {ReactElement, useState} from 'react'
 import type { ReactNode } from 'react'
 import { Header } from '@/components/layout/Header'
-import { Card } from '@/components/ui/card'
-import { Badge } from '@/components/ui/badge'
+import { Badge, Card } from '@/shared/ui'
 import { RoutingRulesPage } from '@/features/routing-rules/RoutingRulesPage'
 import { registryApi, registryKeys } from './api/api'
 import type { InstanceStatus, ServiceVersion, OpenApiDoc } from './api/types'
