+++ service-mesh/frontend/service-mesh/src/features/services/ServicesPage.tsx
@@ -1,10 +1,7 @@
 import { useNavigate } from '@tanstack/react-router'
 import { useQuery } from '@tanstack/react-query'
 import { Header } from '@/components/layout/Header'
-import { Card } from '@/components/ui/card'
-import { Badge } from '@/components/ui/badge'
-import { Skeleton } from '@/components/ui/Skeleton'
-import { ErrorCard } from '@/components/ui/ErrorCard'
+import { Badge, Card, ErrorCard, Skeleton } from '@/shared/ui'
 import { registryApi, registryKeys } from './api/api'
 import type { InstanceStatus } from './api/types'
 import s from './ServicesPage.module.css'
