+++ service-mesh/frontend/service-mesh/src/shared/form/README.md
@@ -1,0 +1,8 @@
+# shared/form
+
+Reserved for the form layer. Will host:
+
+- `createForm.ts` — TanStack Form factory bound to `effect/Schema`
+- `schemaResolver.ts` — bridge from `effect/Schema` decoders to TanStack Form validators
+
+Populated in PR 3 (see `docs/ui-refresh-plan.md`).
