+++ service-mesh/frontend/service-mesh/src/shared/table/README.md
@@ -1,0 +1,7 @@
+# shared/table
+
+Reserved for the data-table layer. Will host:
+
+- `DataTable.tsx` — thin wrapper over TanStack Table that keeps existing CSS Modules
+
+Populated in PR 4 (see `docs/ui-refresh-plan.md`).
