+++ service-mesh/frontend/service-mesh/src/shared/ui/Badge/index.ts
@@ -1,0 +1,1 @@
+export { Badge } from './Badge'
