+++ service-mesh/frontend/service-mesh/src/shared/ui/ErrorCard/index.ts
@@ -1,0 +1,1 @@
+export { ErrorCard } from './ErrorCard'
