+++ service-mesh/frontend/service-mesh/src/shared/ui/Skeleton/Skeleton.tsx
