+++ service-mesh/frontend/service-mesh/src/shared/ui/index.ts
@@ -1,0 +1,8 @@
+// Public barrel for shared UI primitives.
+// Prefer importing from '@/shared/ui' over deep paths.
+
+export { Button } from './Button'
+export { Card, CardHeader, CardTitle, CardValue } from './Card'
+export { Badge } from './Badge'
+export { Skeleton } from './Skeleton'
+export { ErrorCard } from './ErrorCard'
